*Symbol table updates*
-added test code for SymbolTable.java
-added SymbolTable.java based off of RedBlackBST.java (deleted as of
this commit)
-fixed test failure in UniqueSymbolGenerator test caused by improper
generate() function calls
